work_with_datasets/anomaly/requirements.txt
torch
tree-sitter
transformers
tokenizers
GitPython
pickle5
pandas
